chore: move flux lsp files into single dir
